Separei algumas coisas em blueprints e agora os professores são puxados por meio de scraping
