+++ templates/admin/add_events.html
@@ -25,7 +25,7 @@ $(function() {
     <h4>Adicionar evento<h4/>
 </div>
 <div class="panel-body">
-<form method="POST" action="{{ url_for('admin.add_events') }}">
+<form method="POST" action="{{ url_for('crud_events.add_events') }}">
     <div class="form-group">
     {{ form.csrf_token }}
