+++ templates/admin/delete_events.html
@@ -70,7 +70,7 @@ $(function() {
 <button id="edit" type="button" class="btn btn-default">Selecionar</button>
 </div>
 <div class="col-md-10">
-<form method="POST" id="form-edit" action="{{ url_for('admin.delete_events') }}">
+<form method="POST" id="form-edit" action="{{ url_for('crud_events.delete_events') }}">
     <div class="form-group">
     {{ form.csrf_token }}
     {{ form.index(required='true', style="display:none;", readonly="readonly") }}
