+++ templates/admin/edit_events.html
@@ -74,7 +74,7 @@ $(function() {
 <button id="edit" type="button" class="btn btn-default">Selecionar</button>
 </div>
 <div class="col-md-10">
-<form method="POST" id="form-edit" action="{{ url_for('admin.edit_events') }}">
+<form method="POST" id="form-edit" action="{{ url_for('crud_events.edit_events') }}">
     <div class="form-group">
     {{ form.csrf_token }}
     {{ form.index(required='true', style="display:none;", readonly="readonly") }}
