+++ templates/admin/edit_participations.html
@@ -55,7 +55,7 @@ $(document).ready(function() {
 <button id="edit" type="button" class="btn btn-default">Selecionar</button>
 </div>
 <div class="col-md-10">
-<form method="POST" id="form-edit" action="{{ url_for('admin.edit_participations') }}">
+<form method="POST" id="form-edit" action="{{ url_for('crud_participation.edit_participations') }}">
     <div class="form-group">
     {{ form.csrf_token }}
     {{ form.index(required='true', style="display:none;", readonly="readonly") }}
