+++ templates/admin/participations.html
@@ -7,7 +7,7 @@
     <h4>Adicionar intercâmbio<h4/>
 </div>
 <div class="panel-body">
-<form method="POST" action="{{ url_for('admin.participations') }}">
+<form method="POST" action="{{ url_for('crud_participation.participations') }}">
     <div class="form-group">
     {{ form.csrf_token }}
