+++ templates/admin/scheduled_reports.html
@@ -8,7 +8,7 @@
     <h4>Adicionar defesa<h4/>
 </div>
 <div class="panel-body">
-<form method="POST" action="{{ url_for('admin.scheduled_reports') }}">
+<form method="POST" action="{{ url_for('crud_scheduled_reports.scheduled_reports') }}">
     <div class="form-group">
     {{ form.csrf_token }}
