+++ views/admin.py
@@ -113,521 +113,9 @@ def user_loader(user_id):
     else:
         return None
 
-###############################################################################
-#Adicionar deletar e editar defesas de teses e dissertações
-###############################################################################
-
-@APP.route('/apresentacoes/', methods=['GET', 'POST'])
-@login_required
-def scheduled_reports():
-    """
-    Render a report scheduling form.
-    """
-
-    form = ScheduledReportForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.final_reports_dao()
-
-    if form.validate_on_submit():
-        new_report = {
-            'time': form.time.data,
-            'title': form.title.data,
-            'author': form.author.data,
-            'location': form.location.data
-        }
-
-        dao.find_one_and_update(None, {
-            '$push': {'scheduledReports': new_report}
-        })
-
-        return redirect(
-            url_for(
-                'admin.scheduled_reports',
-                success_msg='Defesa de tese adicionada com sucesso.'
-            )
-        )
-
-    return render_template(
-        'admin/scheduled_reports.html',
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/deletar_agendamento/', methods=['GET', 'POST'])
-@login_required
-def delete_scheduled_reports():
-
-    form = ScheduledReportForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.final_reports_dao()
-    json = pfactory.final_reports_dao().find_one()
-    json = dict(json)
-
-    json = dumps(json)
-
-    if form.validate_on_submit() and form.create.data:
-        index = str(form.index.data)
-        dao.find_one_and_update(None, {
-            '$set': {'scheduledReports.' + index + '.deleted' : ""}
-        })
-        return redirect(
-            url_for(
-                'admin.delete_scheduled_reports',
-                final_reports=json,
-                success_msg='Documento deletado com sucesso'
-            )
-        )
-
-    return render_template(
-        'admin/delete_scheduled_reports.html',
-        final_reports=json,
-        form=form,
-        post_graduation=current_user.pg_initials,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/editar_agendamento/', methods=['GET', 'POST'])
-@login_required
-def edit_scheduled_reports():
-
-    form = ScheduledReportForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.final_reports_dao()
-    json = pfactory.final_reports_dao().find_one()
-    json = dict(json)
-    json = dumps(json, ensure_ascii=False)
-
-    if form.validate_on_submit():
-        index = str(form.index.data)
-        new_report = {
-            'time': form.time.data,
-            'title': form.title.data,
-            'author': form.author.data,
-            'location': form.location.data
-        }
-        dao.find_one_and_update(None, {
-            '$set': {'scheduledReports.' + index : new_report}
-        })
-        return redirect(
-            url_for(
-                'admin.edit_scheduled_reports',
-                success_msg='Defesa de tese editada com sucesso.',
-                final_reports=json
-            )
-        )
-
-    return render_template(
-        'admin/edit_scheduled_reports.html',
-        final_reports=json,
-        form=form,
-        post_graduation=current_user.pg_initials,
-        success_msg=request.args.get('success_msg')
-    )
-
 
 ###############################################################################
-#Adicionar deletar e editar membros da equipe de servidores
-###############################################################################
-
-
-
-@APP.route('/add_servidor/', methods=['GET', 'POST'])
-@login_required
-def add_staff():
-    """
-    Render a subject form.
-    """
-
-    form = StaffForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.boards_of_staffs_dao()
-
-    if form.validate_on_submit():
-        if form.photo.data == '':
-            photo = None
-        else:
-            photo = form.photo.data
-        if form.function.data == 'coordination':
-            new_staff = {
-                'name': form.name.data,
-                'rank': form.rank.data,
-                'abstract': form.abstract.data,
-                'photo': photo
-            }
-
-        else:
-            new_staff = {
-                'name': form.name.data,
-                'function': {
-                    'rank': form.rank.data,
-                    'description': form.abstract.data
-                },
-                'photo': photo
-            }
-        dao.find_one_and_update(None, {
-            '$push': {form.function.data: new_staff}
-            })
-        return redirect(
-            url_for(
-                'admin.add_staff',
-                success_msg='Servidor adicionado com sucesso.'
-                )
-        )
-
-    return render_template(
-        'admin/add_staff.html',
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/editar_servidor/', methods=['GET', 'POST'])
-@login_required
-def edit_staff():
-    """
-    Render a subject form.
-    """
-
-    form = StaffForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.boards_of_staffs_dao()
-    json = pfactory.boards_of_staffs_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-
-    dao = pfactory.boards_of_staffs_dao()
-
-    if form.validate_on_submit():
-        index = '.' + str(form.index.data)
-        if form.photo.data == '':
-            photo = None
-        else:
-            photo = form.photo.data
-        if form.function.data == 'coordination':
-            new_staff = {
-                'name': form.name.data,
-                'rank': form.rank.data,
-                'abstract': form.abstract.data,
-                'photo': photo
-            }
-
-        else:
-            new_staff = {
-                'name': form.name.data,
-                'function': {
-                    'rank': form.rank.data,
-                    'description': form.abstract.data
-                },
-                'photo': photo
-            }
-        dao.find_one_and_update(None, {
-            '$set': {form.function.data + index: new_staff}
-            })
-        return redirect(
-            url_for(
-                'admin.edit_staff',
-                success_msg='Servidor editado com sucesso.',
-                staff=json
-                )
-        )
-
-    return render_template(
-        'admin/edit_staff.html',
-        form=form,
-        success_msg=request.args.get('success_msg'),
-        staff=json
-    )
-
-
-@APP.route('/deletar_servidor/', methods=['GET', 'POST'])
-@login_required
-def delete_staff():
-    """
-    Render a delete staff form.
-    """
-
-    form = StaffForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.boards_of_staffs_dao()
-    json = pfactory.boards_of_staffs_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-    index = '.' + str(form.index.data)
-    if form.validate_on_submit():
-        dao.find_one_and_update(None, {
-            '$set': {form.function.data + index + '.deleted' : ''}
-            })
-        return redirect(
-            url_for(
-                'admin.delete_staff',
-                success_msg='Servidor deletado com sucesso.',
-                staff=json
-                )
-        )
-
-    return render_template(
-        'admin/delete_staff.html',
-        form=form,
-        success_msg=request.args.get('success_msg'),
-        staff=json
-    )
-
-###############################################################################
-#Adicionar deletar e editar intercâmbios
-###############################################################################
-
-@APP.route('/intercambios/', methods=['GET', 'POST'])
-@login_required
-def participations():
-    """Render a view for integrations lists."""
-
-    form = ParticipationsInEventsForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.integrations_infos_dao()
-
-    if form.validate_on_submit() and form.create.data:
-        new_participation = {
-            'title': form.title.data,
-            'description': form.description.data,
-            'year': form.year.data,
-            'international': form.location.data
-        }
-
-        dao.find_one_and_update(None, {
-            '$push': {'participationsInEvents': new_participation}
-        })
-
-        return redirect(
-            url_for(
-                'admin.participations',
-                success_msg='Intercâmbio adicionado adicionado com sucesso.'
-            )
-        )
-
-    return render_template(
-        'admin/participations.html',
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/deletar_intercâmbio/', methods=['GET', 'POST'])
-@login_required
-def delete_participations():
-    """
-    Render a delete participation form.
-    """
-
-    form = ParticipationsInEventsForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.integrations_infos_dao()
-    json = pfactory.integrations_infos_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-    index = str(form.index.data)
-
-
-    if form.validate_on_submit():
-        dao.find_one_and_update(None , {
-            '$set': {'participationsInEvents.' + index + '.deleted' : ""}
-        })
-
-        return redirect(
-            url_for(
-                'admin.delete_participations',
-                participations=json,
-                success_msg='Participação deletada com sucesso'
-            )
-        )
-    return render_template(
-        'admin/delete_participations.html',
-        form=form,
-        participations=json,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/editar_intercâmbio/', methods=['GET', 'POST'])
-@login_required
-def edit_participations():
-    """
-    Render a edit participation form.
-    """
-
-    form = ParticipationsInEventsForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.integrations_infos_dao()
-    json = pfactory.integrations_infos_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-    index = str(form.index.data)
-
-    if form.validate_on_submit() and form.create.data:
-        year = int(form.year.data)
-        new_participation = {
-            'title': form.title.data,
-            'description': form.description.data,
-            'year': year,
-            'international': form.location.data
-        }
-
-        dao.find_one_and_update(None, {
-            '$set': {'participationsInEvents.' + index : new_participation}
-        })
-
-        return redirect(
-            url_for(
-                'admin.edit_participations',
-                participations=json,
-                success_msg='Participação editada com sucesso'
-            )
-        )
-    return render_template(
-        'admin/edit_participations.html',
-        form=form,
-        participations=json,
-        success_msg=request.args.get('success_msg')
-    )
-
-###############################################################################
-#Adicionar deletar e editar eventos
-###############################################################################
-
-@APP.route('/add_evento/', methods=['GET', 'POST'])
-@login_required
-def add_events():
-    """Render a view for adding events."""
-
-    form = CalendarForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.calendar_dao()
-    
-    if form.validate_on_submit():
-        initial_date = datetime.datetime.combine(form.initial_date.data, datetime.datetime.min.time())
-        if form.final_date.data != "":
-            final_date = datetime.datetime.strptime(form.final_date.data, '%d/%m/%Y')
-            final_date = datetime.datetime.combine(final_date, datetime.datetime.min.time())
-        else:
-            final_date = form.final_date.data
-        new_event = {
-            'title': form.title.data,
-            'initialDate': initial_date,
-            'finalDate': final_date,
-            'hour': form.hour.data,
-            'link': form.link.data
-        }
-
-        dao.find_one_and_update(None, {
-            '$push': {'events': new_event}
-        })
-
-        return redirect(
-            url_for(
-                'admin.add_events',
-                success_msg='Evento adicionado adicionado com sucesso.'
-            )
-        )
-
-    return render_template(
-        'admin/add_events.html',
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-@APP.route('/editar_evento/', methods=['GET', 'POST'])
-@login_required
-def edit_events():
-    """Render a view for editing events."""
-
-    form = CalendarForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.calendar_dao()
-    json = pfactory.calendar_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-    index = str(form.index.data)
-
-    if form.validate_on_submit() and form.create.data:
-        initial_date = datetime.datetime.combine(form.initial_date.data, datetime.datetime.min.time())
-        if form.final_date.data != "":
-            final_date = datetime.datetime.strptime(form.final_date.data, '%d/%m/%Y')
-            final_date = datetime.datetime.combine(final_date, datetime.datetime.min.time())
-        else:
-            final_date = form.final_date.data
-        new_event = {
-            'title': form.title.data,
-            'initialDate': initial_date,
-            'finalDate': final_date,
-            'hour': form.hour.data,
-            'link': form.link.data
-        }
-
-        dao.find_one_and_update(None, {
-            '$set': {'events.' + index : new_event}
-        })
-
-        return redirect(
-            url_for(
-                'admin.edit_events',
-                events=json,
-                success_msg='Evento editado com sucesso.'
-            )
-        )
-
-    return render_template(
-        'admin/edit_events.html',
-        events=json,
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-
-@APP.route('/deletar_evento/', methods=['GET', 'POST'])
-@login_required
-def delete_events():
-    """Render a view for deleting events."""
-
-    form = CalendarForm()
-
-    pfactory = PosGraduationFactory(current_user.pg_initials)
-    dao = pfactory.calendar_dao()
-    json = pfactory.calendar_dao().find_one()
-    json = dict(json)
-    json = dumps(json)
-
-    if form.validate_on_submit() and form.create.data:
-        index = str(form.index.data)
-        dao.find_one_and_update(None, {
-            '$set': {'events.' + index + '.deleted' : ""}
-        })
-
-        return redirect(
-            url_for(
-                'admin.delete_events',
-                events=json,
-                success_msg='Evento deletado com sucesso.'
-            )
-        )
-
-    return render_template(
-        'admin/delete_events.html',
-        events=json,
-        form=form,
-        success_msg=request.args.get('success_msg')
-    )
-
-
-
-###############################################################################
-#Adicionar deletar e editar professores(Não finalizado)
+#Adicionar deletar e editar professores
 ###############################################################################
 
 
@@ -1013,7 +501,6 @@ def edit_documents():
                 document = form.document.data
                 path = os.path.normpath("static/upload_files/" + current_user.pg_initials.lower())
                 filename = uploadFiles(document, path, document.filename)
-                print(form.category.data, file=sys.stderr)
                 new_document = {
                     'title': form.title.data,
                     'cod': form.cod.data,
